- implemented attach sockets visualization for skeletal mesh (key <A>) - fixed bug from r38 - ClearSkelAnims() was set AnimIndex1 to ANIM_ANASSIGNED (-2), and   CSkelMeshInstance::UpdateSkeleton() was compared AnimIndex1 with INDEX_NONE (-1) only.   As a result, umodel was crashed, when pressing 'Y' twice (start blending animations   test and then stop this test) - support for conversion of Lineage2 FStaticLODModel with rigid sections
- Core:
  - added DO_GUARD option to include/exclude callstack macros guard/unguard
  - included <stddef.h> into Core.h (required for offsetof macro)
  - GlWindow.cpp: changed minimal viewing distance for object from 100 to 50
  - implemented writing of FCompactIndex
  - FFileReader::Serialize() was made public
  - COPY_ARRAY() macro was replaced with CopyArray() function (more compact code)

- genmake: fixed support for source paths like "../dir/file" - was replaced with
  "dir/file" (function AppendFilePath)
- some parts of code was disabled with RENDERING=0 to create GL-less application
- added UT2down test application, which can load UT2 package and downgrade its version
  to 117

git-svn-id: svn://localhost@48 a3f5f06f-15b6-264c-8744-b656ee826ca1

+++ Unreal/UnMesh.h
@@ -887,7 +887,7 @@ struct FSkelMeshSection
 	word			FirstFace;
 	word			NumFaces;
 #if LINEAGE2
-	TArray<int>		LineageBoneMap;
+	TArray<int>		LineageBoneMap;			// for smooth sections only
 #endif
 	// Rigid sections:
 	//	MinWedgeIndex/MaxWedgeIndex -> FStaticLODModel.Wedges
@@ -917,6 +917,7 @@ struct FSkelMeshSection
 
 #if LINEAGE2
 
+// FAnimMeshVertex with influence info
 struct FLineageWedge
 {
 	FVector			Point;
@@ -1203,7 +1204,7 @@ public:
 		if (Ar.IsLineage2)
 		{
 			int unk1, unk3, unk4;
-			TArray<int> unk2;
+			TArray<float> unk2;
 			if (Ar.ArVer >= 118 && Ar.ArLicenseeVer >= 3)
 				Ar << unk1;
 			if (Ar.ArVer >= 123 && Ar.ArLicenseeVer >= 0x12)
